+++ app/src/main/java/com/example/aida/pages/CameraPage.kt
@@ -136,5 +136,6 @@ fun CameraPage(
                         viewModel.startVoiceRecording()
                     })
         )
+
     }
 }
