+++ app/src/main/java/com/example/aida/socketcommunication/VideoClient.kt
@@ -51,27 +51,4 @@ class VideoClient (ip : String = "localhost", port : Int = 12345, timeToTimeout 
             ?.asImageBitmap()
 
     }
-
-    /**
-     * Sends a request to stop the image analysis to the server
-     */
-    fun sendStartGesture(){
-        val id = MessageType.IMAGE_ANALYSIS.value
-        val size = 2
-        val start = Instructions.GESTURE.value
-        val buffer = ByteBuffer.allocate(size)
-        buffer.putShort(start)
-        sendDataToServer(id, buffer.array())
-    }
-    /**
-     * Sends a request to stop the camera feed to the server
-     */
-    fun sendStopGesture() {
-        val id = MessageType.IMAGE_ANALYSIS.value
-        val size = 2
-        val stop = Instructions.OFF.value
-        val buffer = ByteBuffer.allocate(size)
-        buffer.putShort(stop)
-        sendDataToServer(id, buffer.array())
-    }
 }
